- uint32_t instead of int or uint8_t  where applicable - int32_t instead of long where applicable - simplified codeTicksAccessSeq functions - lesser exit sequence points - removed unused LIKELY / UNLIKELY macros (macros for expect) - uint32_t instead of int as return value for armExecute and thumbExecute

+++ trunk/src/vba/gba/Sound.h
@@ -47,8 +47,8 @@ void soundShutdown(void);
 
 //// GBA sound options
 
-long soundGetSampleRate(void);
-void soundSetSampleRate(long sampleRate);
+uint32_t soundGetSampleRate(void);
+void soundSetSampleRate(uint32_t sampleRate);
 
 // Sound settings
